Continuing to build out controller, error from EUtils API
"WebEnv parameter is required" - No idea what this means. According to
the documentation (1) the WebEnv param should only be required when
using Entrez History server. These are not from the History server, so
I need to read up on the docs more. Maybe I should be using the History
server?

1: http://www.ncbi.nlm.nih.gov/books/NBK25499/#_chapter4_EFetch_